# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express Supabase Auth API",
    "version": "1.0.0",
    "description": "API documentation for Express application with Supabase authentication."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthCredentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "example@gmail.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "securepassword123"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Error message details"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "jwt": {
            "type": "string",
            "example": "eyJhbGciOiJFUzI1NiIs..."
          }
        }
      },
      "PublicMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Welcome Stranger, this is public info"
          }
        }
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "description": "Authenticated user data attached by authMiddleware",
            "example": { "id": "145e6104-...", "email": "example@gmail.com" }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentials"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created"
          },
          "400": {
            "description": "Missing email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in with email and password",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns JWT access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out and invalidate session globally",
        "tags": ["Authentication"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logout successful"
          },
          "401": {
            "description": "No token provided or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Get public information",
        "tags": ["Public"],
        "responses": {
          "200": {
            "description": "Successful public response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicMessage"
                }
              }
            }
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get authenticated user profile data",
        "tags": ["Protected"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get protected dashboard access",
        "tags": ["Protected"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard accessed successfully"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          }
        }
      }
    }
  }
}